backend/requirement.txt
flask
requests
python-dotenv
mysql.connector
jwt
